refactor: utility files
