+++ src/pages/blog/category/[category].astro
@@ -3,7 +3,7 @@ import ArrowLeft from '@/assets/icons/arrow-left.svg'
 import ArrowRight from '@/assets/icons/arrow-right.svg'
 import Container from '@/components/Container.astro'
 import PageLayout from '@/layouts/PageLayout.astro'
-import generateOgImageUrl from '@/utils/generateOgImage'
+import generateOgImageUrl from '@/utils/og-image'
 import { getCollection } from 'astro:content'
 
 export async function getStaticPaths() {
@@ -48,7 +48,7 @@ const ogImageUrl = generateOgImageUrl({
             >back to all posts</span
           ></a
         >
-        <h1 class="mb-2 mt-8 text-xl font-bold sm:text-4xl">Blog</h1>
+        <h1 class="mt-8 mb-2 text-xl font-bold sm:text-4xl">Blog</h1>
         <p class="text-foreground-muted sm:text-lg">
           All posts in the <strong>{category}</strong> category
         </p>
@@ -75,7 +75,7 @@ const ogImageUrl = generateOgImageUrl({
                       <li>
                         <a
                           href={`/blog/category/${post.data.category.toLowerCase()}`}
-                          class="rounded-full bg-primary px-2 text-sm text-white hover:bg-primary/80">
+                          class="bg-primary hover:bg-primary/80 rounded-full px-2 text-sm text-white">
                           {post.data.category}
                         </a>
                       </li>
@@ -83,7 +83,7 @@ const ogImageUrl = generateOgImageUrl({
                         <li>
                           <a
                             href={`/blog/tag/${tag.toLowerCase()}`}
-                            class="tag-link rounded-full bg-secondary px-2 text-sm text-black transition-colors hover:text-blue-600">
+                            class="tag-link bg-secondary rounded-full px-2 text-sm text-black transition-colors hover:text-blue-600">
                             {`# ${tag}`}
                           </a>
                         </li>
@@ -91,11 +91,11 @@ const ogImageUrl = generateOgImageUrl({
                     </ul>
                   </div>
                 </div>
-                <div class="hidden text-primary sm:block">
+                <div class="text-primary hidden sm:block">
                   <a href={`/blog/${post.id}?category=${category}`}>
                     <ArrowRight
                       fill="currentColor"
-                      class="mr-0 h-5 w-5 shrink-0 text-primary/50 transition-all duration-200 ease-in-out group-hover:translate-x-2 group-hover:text-blue-600 sm:mr-2"
+                      class="text-primary/50 mr-0 h-5 w-5 shrink-0 transition-all duration-200 ease-in-out group-hover:translate-x-2 group-hover:text-blue-600 sm:mr-2"
                     />
                   </a>
                 </div>
