+++ src/pages/blog/index.astro
@@ -1,10 +1,10 @@
 ---
-import {BlogSearch} from '@/components/BlogSearch.tsx'
+import { BlogSearch } from '@/components/BlogSearch.tsx'
 import Container from '@/components/Container.astro'
 import PageLayout from '@/layouts/PageLayout.astro'
 import { BLOG } from '@/lib/site'
 import { getCollection } from 'astro:content'
-import generateOgImageUrl from '@/utils/generateOgImage'
+import generateOgImageUrl from '@/utils/og-image'
 
 const posts = (await getCollection('blog')).sort(
   (a, b) => b.data.date.valueOf() - a.data.date.valueOf()
