--- src/utils/formatDate.ts
@@ -1,7 +1,0 @@
-export function formatDate(date: string): string {
-  return new Date(date).toLocaleDateString('en-US', {
-    month: 'long',
-    day: '2-digit',
-    year: 'numeric',
-  })
-}
